Update patch set 1
Patch Set 1:

(7 comments)

Thanks for doing it. I was going to add standalone build mode too but you was faster.

 > The error is also happening without this change. Something changed
 > in gerrit master that broke javamelody ide integration but it's not
 > related to this change

That's because gerrit_plugin() rule was broken in this change: [1]. There is also an attempt to fix it pending for review: [2]. Also note, that in my last change: [3] I mentioned that Javamelody plugin is not buildable on current master (as you noticed).

* [1] https://gerrit-review.googlesource.com/67043
* [2] https://gerrit-review.googlesource.com/67186
* [3] https://gerrit-review.googlesource.com/67408

Patch-set: 1
Label: Code-Review=0