Fix extension-api dependency
The build [1] is failing with the error (wrapped for
readability):

  Couldn't get dependency '//lib/gerrit:extension-api' of target
  '//plugins/javamelody:javamelody-datasource-interceptor-lib':

Update the dependency to fix the build and align with Gerrit stable-2.11
and at the same time makes an equivalent definition for the standalone build.

Please note that the corresponding fix on Gerrit master is going
to be different as they will be based on a different version of bucklets.

[1] https://gerrit-ci.gerritforge.com/job/plugin-javamelody-stable-2.11/8/console

Change-Id: I57702e7f3d3a9501d2e74f7e800bacf2e4b47f74