Update patch set 1
Patch Set 1:

There is a fundamental problem with the standalone build: it cannot work on master at all as the Gerrit version isn't published on Maven.

I would keep this fix for master because it is correct and standalone build is not applicable.

I will then fix the standalone build on stable-2.11

Patch-set: 1