Update patch set 2
Patch Set 2:

> Not really, for stable-2.12 (new branch to create) we should switch
 > to 2.12-rc0, master is 2.13-SNAPSHOT instead.

On master update to 2.12-rc0, cut stable-2.12 from that, and then on master update to 2.13-SNAPSHOT, and then we have a history like this:

 O (master) update to 2.13-SNAPSHOT
 |
 O (stable-2.12) update to 2.12-rc0
 |
 O ...

This is how I've done it for recent updates to 2.12-rc0, although I didn't also do the 2.13-SNAPSHOT update yet for all cases.

By branching out first, we create an unnecessary fork and merge:

 O (master) update to 2.13-SNAPSHOT
 O (master) merge branch stable-2.12
 | \
 | |
 | O (stable-2.12) update to 2.12-rc0
 |/
 O ...

In the previous scenario we only create a fork when (or if) we need to make fixes specific for stable-2.12.

Patch-set: 2
Reviewer: Gerrit User 1011123 <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Verified=0