Include needed dependencies in javamelody.jar
So far, when building the plugin as stated in the documentation, the
resulting jar did not include the needed libraries for the plugin to
load in Gerrit as they were packaged in the javamelody-deps_deploy.jar.

Deploying the javamelody-deps_deploy.jar in the gerrit site lib folder
fixes the problem but that means the behaviour of the bazel build is
different from the one in buck and this is not documented.

Adding the required dependencies to the gerrit_plugin rule allows to
build a javamelody jar that behaves as other plugins (i.e., it suffices
to copy it to the plugins folder) and that is almost identical (except
for the order of some entries in the manifest) to the one generated by
buck.

If database monitoring capabilities are desired, deploying the (renamed)
javamelody-datasource_deploy.jar to the lib folder in the gerrit site
and adding the corresponding configuration to the gerrit.config file
activates this feature.

Test plan:

1. Build javamelody.jar and deploy it to plugins folder
2. Verify Monitoring tab is shown and it is functional
3. Build javamelody-datasource_deploy.jar and deploy it to lib folder
4. Verify no database statistics are shown in Monitoring page
5. Add configuration to gerrit.config to activate datasource
   interception
6. Verify database statistics are now shown in Monitoring page

Change-Id: Id294d1bf9a5f0f9b6b831f3e2620616f8899b7dc