Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> @Luca, What about Prometheus integration without Javamelody, but using Dropwizard metrics in gerrit core?
> 
> The idea would be to write dedicated Prometheus plugin that would expose Dropwizard metrics in Prometheus format, so the machinery would be something like:
> 
> Gerrit core + Gerrit-Prometheus-plugin + Prometheus/Dropwizard Servlet | <= Prometheus => Grafana Dashboard / Prometheus Dashboard. Advantage is clear: Gerrit core Dropwizard metrics are more specific to Gerrit internals compared to generic Javamelody metrics.
> 
> Searching for Dropwizard and Prometheus integration reveals number of approaches, e.g.: [1].
> 
> @David/Gustaf/Sven: have you considered (or even implemented) something like that already?
> 
> [1] https://www.robustperception.io/exposing-dropwizard-metrics-to-prometheus

I'm not that familiar with Prometheus but I know it likes the http polling approach for fetching data. Not sure how it would affect Prometheus, but to me it seems reasonable to use the metrics-reporter-prometheus plugin for the actual writing of data to Prometheus and then utilizing Gerrit's plugin metrics support (just like the replication plugin does) for supplying the data. This way all the other plugins can make use of the data JMelody provides.

Patch-set: 7