Allow access to metrics from Prometheus
Enable the remote access to metrics scrubber from Prometheus
using a bearer token.

Currently, this is the only way to scrap metrics through the JavaMelody
filter because basic auth isn't an option yet with Gerrit, because
the AllRequestsFilter is invoked well before the basic auth filter.

Enabled only when the request URL matches the Prometheus
scraping query, otherwise everything is delegated to the normal
behaviour of JavaMelody and the associated access controls.

Change-Id: I2312a4e225ce1e5b25f46414a1c8f0625dd8842f