Allow access using Basic Authentication prefix
Enable the use of '/a/' prefix to the monitoring URL to allow external
authenticated access to the monitoring data (e.g. Prometheus
integration).

Example:
  curl -u user:secret http://gerrit/a/monitoring?format=prometheus

One problem is that current Javamelody core doesn't support custom
prefixes for monitoring URL detection. This PR fixes that: [1], that
wasn't merged yet. Consume Javamelody core built with this PR included.

Another problem is that use authentication currently is broken for
filters that utilize AllRequestFilter. That is fixed in: [2].

With these two fixes in place, Prometheus scraping query just works.

[1] https://github.com/javamelody/javamelody/pull/766
[3] https://gerrit-review.googlesource.com/c/gerrit/+/192430

Bug: Issue 9550
Change-Id: Iecfe4fd2ea79fb19c02a81b181e78df01f73cd1b