Allow access using Basic Authentication prefix
Enable the use of '/a/' prefix to the monitoring URL to allow external
authenticated access to the monitoring data (e.g. Prometheus
integration):

  GET /a/monitoring?format=prometheus

One problem is that current Javamelody core doesn't support custom
prefixes for monitoring URL detection. This PR fixes that: [1], that
wasn't merged yet. Consume Javamelody core built with this PR included.

Another problem is that user authentication is currently broken for
filters that relies on AllRequestFilter. That CL fixes that: [2].

With these two fixes in place, Prometheus scraping query just works,
and that without changing any line of code in Javamelody plugin itself.

One advantage of this approach is that this allows scraping of two kinds
of mertrics: Gerrit core own (exposed by Dropwizard) and Javamelody own
metrics, both with Basic Authentication, in combination with
metrics-reporter-prometheus, using this endpoint:

  GET /a/plugins/metrics-reporter-prometheus/metrics

To consume custom javamelody-core artifact, java_import_external rule is
used. It appears, that this rule is partly broken on most recent Bazel
master. That is fixed in this PR: [3].

[1] https://github.com/javamelody/javamelody/pull/766
[2] https://gerrit-review.googlesource.com/c/gerrit/+/192430
[3] https://github.com/bazelbuild/bazel/pull/5864

Bug: Issue 9550
Change-Id: Iecfe4fd2ea79fb19c02a81b181e78df01f73cd1b