Update patch set 1
Patch Set 1:

As you know, both sites: OpenStack and Libreoffice are still on 2.13. Even if it  is not cLear yet whether they would skip 2.14 or not I prefer to be prepared with usage of this plugin. But we should certainly cherry-pick your change.

Patch-set: 1
Reviewer: Gerrit User 1011323 <1011323@173816e5-2b9a-37c3-8a2e-48639d4f1153>