Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+2
> 
> Thanks. Note, that on master, where ReviewDb gone, and jdbc interceptor is not needed anymore, the build and deployment of javamelody plugin is significantly simplified.

yep

since last Saturday we finally have all servers on stable-2.16, though there is still some way to go. First we need to migrate our huge sites to noteDB. We are working on doing the bulk of this in a replica and then replicate the git notes resulting from the migration in the replica back to the master so that we can do the migration of the delta in an acceptable downtime.

Meanwhile we mostly rely on gerrit-monitoring running on k8s, only for a few javamelody features javamelody we still don't have a solution in gerrit-monitoring. E.g. a quick way to get a thread dump without having to open a ssh connection to the server.

Patch-set: 1